move : member용 dto로 파일 이동
